requirements.txt
numpy>=1.26
pandas>=2.1
pyarrow>=14
orjson>=3.9
tqdm>=4.66
pyyaml>=6.0
scikit-learn>=1.3
sentence-transformers>=2.6
faiss-cpu>=1.8
python-igraph>=0.11
leidenalg>=0.10